Move rating and notifications prompt to done page.
